TestCases/requirements.txt
selenium
pytest
allure-pytest
pytest-html
pytest-soft-assertions
pytest-dependency
pytest-order
chromedriver-autoinstaller
openpyxl